attack
created attack and take damage functions for ship script and
usershipagent.  If user presses Q or E the cannon from the corresponding
side is 'shot'.  A raycast is used to check for a ship within range and
implements damage on that ship.  Enemy ships must be tagged 'Enemy' for
this to work.