Wood Buttons and Upgrade Numbers
I changed the buttons in the Shop GUI to have the wood texture, plus
they are a bit more spaced out now. I tried to get numbers set up to
show how many of a particular upgrade the player had purchased so far.
The number do update in accordance with the number of upgrades, but I'm
having issues with getting the numbers to display correctly next to the
buttons at different screen sizes. Also, even though I have it set up in
GUI Functions to have the numbers disappear if the Shop GUI is not
active, they still show up after closing the Shop.  I will try again
tomorrow to fix this issue.  If this issue cannot be fixed, I think it
might be best just to drop that feature, maybe have a separate Pause
menu that the Player can open to look at their stats and upgrades.