Count added to GUIFunctions
I tried to get it so that the text within the button itself changes on
each upgrade ("Cannons: 0" to "Cannons: 1" and so on), but I could not
get it to work. Apparently, it is possible to change the text in a GUI
button such that there is a variable attached, and the text updates
along with the variable. However, this requires that the button is
created through scripting only rather than by the canvas that is used
for the buttons we have now. This causes a problem because much like the
menuText that is in GUIFunctions right now, this would mean that the
button will not be able to stay in the same relative spot if the screen
is resized.

There is a similar problem with trying to make a separate "Check stats"
page like I mentioned yesterday. We could have the variables appear
within a standalone GUIText on a stats screen through scripts, but there
would still be the same problem with positioning the numbers in the same
spot when the screen is resized.

I noticed that for each button's text, in the inspector it labels it as
"Text (Script)". I thought this meant that the text has its own script,
and I could add in the updating numbers by changing its own script.
However, when I right click on Text (Script)'s gear in the right side
corner, and click on Edit Script from the menu, it doesn't open a script
in MonoDevelop. I am not sure if this is my computer acting up, or if
these scripts cannot be edited by default.

For now, I left in the counting variables for each upgrade in
GUIFunctions and made them public. That way, at least we can check that
the variables are updating within the Unity inspector, but not the game
itself. We might be able to find a solution to this problem later on.
However, from what I have read and what I understand about Unity's GUI
system, this might require making an entirely new GUI through scripts
only and would cause positioning problems whenever the game screen is
resized.