cannon and death animations
animated the cannons shooting from ship to end of shooting range.  If
another ship is hit then the explode animation is triggered too.
Implemented the firing rate.
When a ship's health=0 it is respawned.  The ship state is changed to
ship.dying so that it cannot do anything while respawning.  The ship
flashes where is dies, then reappears in its harbor(facing the island)
and flashes some more.  Once the flashing stops its state changes back
to roaming.  Health is reset back to 5.