GUI fixes
I added "Each upgrade costs 100 coins" to the GUI, and also fixed the
if-statements in GUIFunctions so that the player will only be able to
get 2 of a particular kind of upgrade. I tried to do what Sarah
suggested for the  GUI text change, but for whatever reason it doesn't
seem to be working. You can see the new sections of code in
GUIFunctions.cs, they all have the comment "//Code for changing the
button text." I made sure the button text in the ShopGUI were renamed
accordingly, and I also tried this code with text objects that are
separate from the buttons, but neither of them seem to work. At the
least, they were not updating at all.

I also tried the code in both cases with no text in the text objects (So
where it would say 'New Text' by default in the Inspector, I made sure
it was left blank), and that didn't work either. I made sure to follow
the code in GoldScript.cs, so I am not sure what I did wrong.

(On that note, I found out that GoldScript.cs isn't in the project
anymore. The only way I could look at the code was to check the history
of the project in the Master branch. The merge that says "Merge branch
'maxhealth'" shows GoldScript.cs, but it shows all the code in red.
Doesn't that mean that it was deleted? Just thought I should let you
know.)